completely decoupling test suite from Lightbeam class per suggestion from John
